Changed sign-in & sign-up alert to use global alert
