+++ src/state/alert/dispatchers.ts
@@ -1,21 +1,9 @@
-import {PayloadAction} from '@reduxjs/toolkit';
+import {createAction} from '@reduxjs/toolkit';
 
 import {ShowAlertData} from './data';
-import {AlertState} from './state';
-
-export enum AlertDispatcherName {
-  SHOW_ALERT = 'showAlert',
-  ALERT_CLOSED = 'alertClosed',
-}
+import {AlertDispatcherName} from './name';
 
 export const alertDispatchers = {
-  [AlertDispatcherName.SHOW_ALERT]: (state: AlertState, {payload}: PayloadAction<ShowAlertData>) => {
-    state.message = payload.message;
-    state.severity = payload.severity;
-    state.showAlert = true;
-  },
-  [AlertDispatcherName.ALERT_CLOSED]: (state: AlertState) => {
-    state.message = '';
-    state.showAlert = false;
-  },
+  [AlertDispatcherName.SHOW_ALERT]: createAction<ShowAlertData>(AlertDispatcherName.SHOW_ALERT),
+  [AlertDispatcherName.ALERT_CLOSED]: createAction(AlertDispatcherName.ALERT_CLOSED),
 };
