+++ src/state/alert/name.ts
@@ -1,0 +1,6 @@
+export const ALERT_STATE_NAME = 'alert';
+
+export enum AlertDispatcherName {
+  SHOW_ALERT = 'showAlert',
+  ALERT_CLOSED = 'alertClosed',
+}
