+++ src/state/auth/name.ts
@@ -1,0 +1,8 @@
+export const AUTH_STATE_NAME = 'auth';
+
+// noinspection JSUnusedGlobalSymbols
+export enum AuthDispatcherName {
+  SIGN_UP = 'signUp',
+  SIGN_IN = 'signIn',
+  SIGN_OUT = 'signOut',
+}
